Import greylist.pl from prod

+++ greylist.pl
@@ -6,8 +6,10 @@
 use strict;
 use warnings;
 
+use Carp;
 use DBI;
 use DBD::SQLite;
+use Domain::PublicSuffix;
 use Time::Local;
 use Net::Netmask;
 
@@ -24,14 +26,43 @@ my $output = "E451 Greylisting active. Your mail will be accepted on the next at
 # By default, greylisting is active
 my $greylist = 1;
 
+# Handle IPv6
+my $tcpremoteip;
+if ($ENV{'PROTO'} eq 'TCP')  { $tcpremoteip = $ENV{'TCPREMOTEIP'} }
+if ($ENV{'PROTO'} eq 'TCP6') { $tcpremoteip = "0.0.0.0" }
+
+# Make sure TCPREMOTEHOST is set
+my $tcpremotehost = "";
+if (defined($ENV{'TCPREMOTEHOST'})) {
+  $tcpremotehost = $ENV{'TCPREMOTEHOST'};
+}
+if (defined($ENV{'TCP6REMOTEHOST'})) {
+  $tcpremotehost = $ENV{'TCP6REMOTEHOST'};
+}
+
 # Net::Netmask object
-my $block = new Net::Netmask ($ENV{'TCPREMOTEIP'}, $mask);
+my $block = Net::Netmask->new($tcpremoteip, $mask);
 my $network = $block->base();
 
+# Domain::PublicSuffix object
+my $suffix = Domain::PublicSuffix->new;
+my $rdns;
+if (defined($suffix->get_root_domain($tcpremotehost))) {
+  $rdns = "".$suffix->get_root_domain($tcpremotehost);
+} else {
+  $rdns = "";
+}
+
 # Relayclients will not be greylisted, of course.
-if ( (defined($ENV{'RELAYCLIENT'})) || 
-     (defined($ENV{'TRUSTCLIENT'})) || 
+if ( (defined($ENV{'RELAYCLIENT'})) ||
+     (defined($ENV{'TRUSTCLIENT'})) ||
      (not(-w "/var/qmail/control/greylist/")) ) {
+  print "\n";
+  exit(0);
+}
+
+# hazel hack
+if ($ENV{'SMTPRCPTTO'} =~ /hetzl\.com/ix) {
   print "\n";
   exit(0);
 }
@@ -44,16 +75,18 @@ if (-f "/var/qmail/control/greylist/disable") {
 
 my $i;
 my $now = timelocal(localtime);
-$ENV{'SMTPMAILFROM'} =~ /\@(.*)$/;
-my $sdomain = $1;
+my $sdomain = "";
+if ($ENV{'SMTPMAILFROM'} =~ /\@(.*)$/x) {
+  $sdomain = "$1";
+}
 
-my $dbh = DBI->connect("dbi:SQLite:dbname=/var/qmail/control/greylist/sqlite.db","","") 
-          || die "Cannot connect: $DBI::errstr";
+my $dbh = DBI->connect("dbi:SQLite:dbname=/var/qmail/control/greylist/sqlite.db","","")
+          || croak "Cannot connect: $DBI::errstr";
 
 my $res = $dbh->selectall_arrayref( "SELECT timestamp FROM main WHERE ipaddr = \"$network\" AND mail = \"$sdomain\" AND rcpt = \"$ENV{'SMTPRCPTTO'}\"" );
 
-foreach( @$res ) {
-  foreach $i (0..$#$_) {
+foreach (@$res) {
+  foreach my $i (0..$#$_) {
     # Check if the 'triplet' is older then $mindelay and younger then $maxvalid
     if ( ($_->[$i] >= ($now - $maxvalid)) && ($_->[$i] <= ($now - $mindelay)) ) {
       $greylist = 0;
@@ -63,14 +96,17 @@ foreach( @$res ) {
 }
 
 # Record the current 'triplet' to allow next attempt
-$dbh->do( "INSERT INTO main VALUES (\"$network\",\"$sdomain\",\"$ENV{'SMTPRCPTTO'}\",\"$now\")" );
+$dbh->do( "INSERT INTO main VALUES (\"$network\",\"$sdomain\",\"$ENV{'SMTPRCPTTO'}\",\"$now\",\"$rdns\")" );
 
 # Delete old entries in the DB (older than $maxvalid) and clean out the DB
-$dbh->do( "DELETE FROM main WHERE timestamp < $now - $maxvalid" );
-$dbh->do( "VACUUM" );
+if (($$ % 100) == 0) {
+  print STDERR getppid()." Vacuuming greylist database\n";
+  $dbh->do( "DELETE FROM main WHERE timestamp < $now - $maxvalid" );
+  $dbh->do( "VACUUM" );
+}
 
-if ($greylist) { 
-  print STDERR getppid()." Added to greylist\n"; 
+if ($greylist) {
+  print STDERR getppid()." Added to greylist\n";
   print $output;
 } else {
   print STDERR getppid()." Passed greylist test\n";
